requirements.txt
google-cloud-vision==2.1.0

google-cloud-bigquery==1.27.2
